+++ service/service.go
@@ -9,7 +9,7 @@ type AccountService interface {
 
 type TransactionService interface {
 	Create(transaction *entity.Transaction) error
-	Find(id int32) (entity.Transaction, error)
+	Find(id int64) (entity.Transaction, error)
 	FindByAccountID(id int64) ([]entity.Transaction, error)
 	Update(transaction *entity.Transaction) error
 	UpdateBalance(transaction *entity.Transaction) error
